Timeout for directory removal, support ingen script, always make zip skeleton
Without timeout the process freezed. ingen script is specified as the `in` parameter and is expected to write tests like `xyz3b.in` in the current working directory when ran. The skeleton is xyz/doc/, xyz/prog/, xyz/in/, xyz/out/ and if it doesn't exist the package type is not recognized.